#632 Apply PR feedback: update localized inputs, types, tests and docs
