+++ packages/metahubs-backend/base/src/routes/metahubsRoutes.ts
@@ -16,6 +16,7 @@ import { validateListQuery } from '../schemas/queryParams'
 import { createLocalizedContent, updateLocalizedContentLocale } from '@universo/utils'
 import { escapeLikeWildcards } from '../utils'
 import type { VersionedLocalizedContent } from '@universo/types'
+import { isValidLocaleCode } from '@universo/types'
 
 /**
  * Get the appropriate manager for the request (RLS-enabled if available)
@@ -34,9 +35,14 @@ const resolveUserId = (req: Request): string | undefined => {
 const sanitizeLocalizedInput = (input: Record<string, string | undefined>) => {
     const sanitized: Record<string, string> = {}
     for (const [locale, value] of Object.entries(input)) {
-        if (typeof value === 'string' && value.trim() !== '') {
-            sanitized[locale] = value.trim()
-        }
+        if (typeof value !== 'string') continue
+        const trimmedValue = value.trim()
+        if (!trimmedValue) continue
+        const normalized = locale.trim().replace('_', '-')
+        const [lang, region] = normalized.split('-')
+        const normalizedCode = region ? `${lang.toLowerCase()}-${region.toUpperCase()}` : lang.toLowerCase()
+        if (!isValidLocaleCode(normalizedCode)) continue
+        sanitized[normalizedCode] = trimmedValue
     }
     return sanitized
 }
@@ -46,15 +52,17 @@ const buildLocalizedContent = (
     primaryLocale?: string,
     fallbackPrimary?: string
 ): VersionedLocalizedContent<string> | undefined => {
-    const entries = Object.entries(input)
-    if (entries.length === 0) return undefined
+    const localeCodes = Object.keys(input).sort()
+    if (localeCodes.length === 0) return undefined
 
     const primaryCandidate = primaryLocale && input[primaryLocale] ? primaryLocale : fallbackPrimary && input[fallbackPrimary] ? fallbackPrimary : undefined
-    const primary = primaryCandidate ?? entries[0][0]
+    const primary = primaryCandidate ?? localeCodes[0]
     let content = createLocalizedContent(primary, input[primary] ?? '')
 
-    for (const [locale, value] of entries) {
+    for (const locale of localeCodes) {
         if (locale === primary) continue
+        const value = input[locale]
+        if (typeof value !== 'string') continue
         content = updateLocalizedContentLocale(content, locale, value)
     }
