+++ packages/metahubs-frontend/base/src/utils/localizedInput.ts
@@ -4,6 +4,16 @@ import type { SimpleLocalizedInput } from '../types'
 
 export const normalizeLocale = (locale?: string) => (locale ? locale.split(/[-_]/)[0].toLowerCase() : 'en')
 
+export const getSimpleLocalizedValue = (field: SimpleLocalizedInput, locale = 'en'): string => {
+    const normalized = normalizeLocale(locale)
+    const localized = field[normalized]
+    if (typeof localized === 'string' && localized.trim() !== '') return localized
+    if (typeof field.en === 'string' && field.en.trim() !== '') return field.en
+    if (typeof field.ru === 'string' && field.ru.trim() !== '') return field.ru
+    const fallback = Object.values(field).find((value) => typeof value === 'string' && value.trim() !== '')
+    return typeof fallback === 'string' ? fallback : ''
+}
+
 export const extractLocalizedInput = (value?: VersionedLocalizedContent<string> | null) => {
     const filtered = filterLocalizedContent(value)
     if (!filtered) return { input: undefined, primaryLocale: undefined }
